+++ main.js
@@ -1,96 +1,14 @@
-const $arenas = document.querySelector('.arenas');
+// Подключаем функции и переменные
+import { player1, player2, showResult, enemyAttack, playerAttack } from './fight.js';
+import {createElement} from './utils.js';
+import { generateLogs, getLogMessage } from './fight_logs.js';
+
+export const $arenas = document.querySelector('.arenas');
 // const $randButton = document.querySelector('.button');
-const $fightForm = document.querySelector('.control');
-const $chat = document.querySelector('.chat');
+export const $fightForm = document.querySelector('.control');
 
-const player1 = {
-    player: 1,
-    name: 'SUBZERO',
-    hp: 100,
-    img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif',
-    weapon: [
-        'sword',
-        'shuriken',
-        'kunai'
-    ],
-    attack
-};
 
-const player2 = {
-    player: 2,
-    name: 'LIU KANG',
-    hp: 100,
-    img: 'http://reactmarathon-api.herokuapp.com/assets/liukang.gif',
-    weapon: [
-        'sword',
-        'shuriken',
-        'kunai'
-    ],
-    attack
-};
 
-const HIT = {
-    head: 30,
-    body: 25,
-    foot: 20,
-}
-const ATTACK = [
-    'head',
-    'body',
-    'foot'
-];
-
-const logs = {
-    start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
-    end: [
-        'Результат удара [playerWins]: [playerLose] - труп',
-        '[playerLose] погиб от удара бойца [playerWins]',
-        'Результат боя: [playerLose] - жертва, [playerWins] - убийца',
-    ],
-    hit: [
-        '[playerDefence] пытался сконцентрироваться, но [playerKick] разбежавшись раздробил копчиком левое ухо врага.',
-        '[playerDefence] расстроился, как вдруг, неожиданно [playerKick] случайно раздробил грудью грудину противника.',
-        '[playerDefence] зажмурился, а в это время [playerKick], прослезившись, раздробил кулаком пах оппонента.',
-        '[playerDefence] чесал <вырезано цензурой>, и внезапно неустрашимый [playerKick] отчаянно размозжил грудью левый бицепс оппонента.',
-        '[playerDefence] задумался, но внезапно [playerKick] случайно влепил грубый удар копчиком в пояс оппонента.',
-        '[playerDefence] ковырялся в зубах, но [playerKick] проснувшись влепил тяжелый удар пальцем в кадык врага.',
-        '[playerDefence] вспомнил что-то важное, но внезапно [playerKick] зевнув, размозжил открытой ладонью челюсть противника.',
-        '[playerDefence] осмотрелся, и в это время [playerKick] мимоходом раздробил стопой аппендикс соперника.',
-        '[playerDefence] кашлянул, но внезапно [playerKick] показав палец, размозжил пальцем грудь соперника.',
-        '[playerDefence] пытался что-то сказать, а жестокий [playerKick] проснувшись размозжил копчиком левую ногу противника.',
-        '[playerDefence] забылся, как внезапно безумный [playerKick] со скуки, влепил удар коленом в левый бок соперника.',
-        '[playerDefence] поперхнулся, а за это [playerKick] мимоходом раздробил коленом висок врага.',
-        '[playerDefence] расстроился, а в это время наглый [playerKick] пошатнувшись размозжил копчиком губы оппонента.',
-        '[playerDefence] осмотрелся, но внезапно [playerKick] робко размозжил коленом левый глаз противника.',
-        '[playerDefence] осмотрелся, а [playerKick] вломил дробящий удар плечом, пробив блок, куда обычно не бьют оппонента.',
-        '[playerDefence] ковырялся в зубах, как вдруг, неожиданно [playerKick] отчаянно размозжил плечом мышцы пресса оппонента.',
-        '[playerDefence] пришел в себя, и в это время [playerKick] провел разбивающий удар кистью руки, пробив блок, в голень противника.',
-        '[playerDefence] пошатнулся, а в это время [playerKick] хихикая влепил грубый удар открытой ладонью по бедрам врага.',
-    ],
-    defence: [
-        '[playerKick] потерял момент и храбрый [playerDefence] отпрыгнул от удара открытой ладонью в ключицу.',
-        '[playerKick] не контролировал ситуацию, и потому [playerDefence] поставил блок на удар пяткой в правую грудь.',
-        '[playerKick] потерял момент и [playerDefence] поставил блок на удар коленом по селезенке.',
-        '[playerKick] поскользнулся и задумчивый [playerDefence] поставил блок на тычок головой в бровь.',
-        '[playerKick] старался провести удар, но непобедимый [playerDefence] ушел в сторону от удара копчиком прямо в пятку.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.',
-        '[playerKick] не думал о бое, потому расстроенный [playerDefence] отпрыгнул от удара кулаком куда обычно не бьют.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.'
-    ],
-    draw: 'Ничья - это тоже победа!'
-};
-
-function attack() {
-    return console.log('Fight...' + this.name);
-};
-
-function createElement(tag, className) {
-    const $tag = document.createElement(tag);
-    if (className) {
-        $tag.classList.add(className);
-    }
-    return $tag;
-};
 
 function createPlayer(playerObj) {
     const $player = createElement('div', `player${playerObj.player}`);
@@ -115,216 +33,8 @@ function createPlayer(playerObj) {
     return $player;
 };
 
-// function getRandom(num) {
-//     // Получаем силу удара
-//     const randNum = Math.ceil(Math.random() * num);
-//     return randNum;
-// }
-
-const getRandom = (num) => Math.ceil(Math.random() * num);
-
-Object.prototype.changeHP = function(hit) {
-    // Отнимаем здоровье
-    this.hp -= hit;
-
-    // Если здоровье 0 или меньше устанавливаем hp = 0
-    if (this.hp <= 0 ) { 
-        this.hp = 0;
-    };
-};
-
-Object.prototype.elHP = function() {
-    // Получаем блок здоровья игрока
-    const $player = document.querySelector(`.player${this.player} .life`);
-    
-    return $player;
-};
-
-// const elHP = () => document.querySelector(`.player${this.player} .life`);
-
-Object.prototype.renderHP = function() {
-    // Отображаем изменение шкалы здоровья
-    const $playerLife = this.elHP();
-
-    $playerLife.style.width = `${this.hp}%`;
-};
-
-// Функция получает имя победителя и создает блок с именем победителя 
-function playerWin(name) {
-    const $winTitle = createElement('div', 'loseTitle');
-    if (name) {
-        $winTitle.innerText = `${name} WINS!`;
-    } else {
-        // Выводим блок с сообщением о ничьей
-        $winTitle.innerText = 'DOUBLE KILL!';
-    };
-
-    return $winTitle;
-};
-
-function getWinner() {
-    if (player1.hp === 0 && player1.hp < player2.hp) {
-        // Если игрок 1 проиграл выводим имя 2 игрока
-        $arenas.appendChild(playerWin(player2.name));
-        // generateLogs('end', player2.name, player1.name);
-        generateLogs(getLogMessage('end', player2.name, player1.name));
-    } else if (player2.hp === 0 && player2.hp < player1.hp) {
-        // Если игрок 2 проиграл выводим имя 1 игрока
-        $arenas.appendChild(playerWin(player1.name));
-        // generateLogs('end', player1.name, player2.name);
-        generateLogs(getLogMessage('end', player1.name, player2.name));
-    } else if (player1.hp === 0 && player2.hp === 0) {
-        // Если ничья выводим сообшение double kill
-        $arenas.appendChild(playerWin());
-        generateLogs(getLogMessage('draw'));
-    };
-};
-
-function createReloadButton() {
-    // Создаем кконтейнер и кнопку перезагрузки
-    const $reloadWrap = createElement('div', 'reloadWrap');
-    const $reloadButton = createElement('button', 'button');
-
-    $reloadButton.innerText = 'Restart';
-
-    // Добавляем кнопку в DOM-дерево
-    $reloadWrap.appendChild($reloadButton);
-    $arenas.appendChild($reloadWrap);
-
-    // Обрабатываем клик по кнопке
-    $reloadButton.addEventListener('click', function() {
-        // Обновляем страницу
-        window.location.reload();
-    });
-};
-
-// $randButton.addEventListener('click', function() {
-//     // Вызываем метод изменения здоровья
-//     player1.changeHP(getRandom(20));
-//     // Вызываем метод отображения здоровья
-//     player1.renderHP();
-
-//     player2.changeHP(getRandom(20))
-//     player2.renderHP();
-    
-//     if (player1.hp === 0 || player2.hp === 0) {
-//         // Отлючаем кнопку
-//         $randButton.disabled = true;
-//         getWinner();
-//         // Вызываем фунцию создания кнопки перезагрузки
-//         createReloadButton();
-//     }
-// });
-
-function showResult() {
-    if (player1.hp === 0 || player2.hp === 0) {
-        // Отлючаем кнопку и выбор действий
-        for (let item of $fightForm) {
-
-            if (item.type === 'submit' || item.type === 'radio') {
-                item.disabled = true;
-            }
-        }
-        // Выводим победителя
-        getWinner();
-        // Вызываем фунцию создания кнопки перезагрузки
-        createReloadButton();
-    }
-};
 
 
-function enemyAttack() {
-    const enemyHit = ATTACK[getRandom(3) - 1];
-    const enemyDefence = ATTACK[getRandom(3) - 1];
-
-    return {
-        enemyHitValue: getRandom(HIT[enemyHit]),
-        enemyHit,
-        enemyDefence,
-    }
-};
-
-function playerAttack() {
-    const attack = {};
-
-    for (let item of $fightForm) {
-        if (item.checked && item.name === 'hit') {
-            attack.playerHitValue = getRandom(HIT[item.value]);
-            attack.playerHit = item.value;
-        }
-
-        if (item.checked && item.name === 'defence') {
-            attack.playerDefence = item.value;
-        }
-
-        item.checked = false;
-    };
-
-    return attack;
-
-};
-
-// function getTime() {
-//     const time = new Date();
-//     const normalize = (num) => (num.toString().length > 1 ? num : `0${num}`);
-//     const fightTime = `${normalize(time.getHours())}:${normalize(time.getMinutes())}:${normalize(time.getSeconds())}`;
-
-//     return fightTime;
-// };
-
-const getTime = () => {
-    const time = new Date();
-    const normalize = (num) => (num.toString().length > 1 ? num : `0${num}`);
-    const fightTime = `${normalize(time.getHours())}:${normalize(time.getMinutes())}:${normalize(time.getSeconds())}`;
-
-    return fightTime;
-};
-
-function getLogMessage(type, player1, player2) {
-    
-    let text;
-    switch(type) {
-        case 'start' :
-            text = logs[type]
-                .replace('[player1]', player1)
-                .replace('[player2]', player2);
-            return text;
-        case 'hit' :
-            text = logs[type][getRandom(logs[type].length) - 1]
-                .replace('[playerKick]', player1)
-                .replace('[playerDefence]', player2);
-            return text;
-        case 'defence' :
-            text = logs[type][getRandom(logs[type].length) - 1]
-                .replace('[playerKick]', player1)
-                .replace('[playerDefence]', player2);
-            return text;
-        case 'end' :
-            text = logs[type][getRandom(logs[type].length) - 1]
-                .replace('[playerWins]', player1)
-                .replace('[playerLose]', player2);
-            return text;
-        case 'draw' :
-            text = logs[type];
-            return text;
-        default :
-            text = '... ждем ...';
-            return text;
-    }
-};
-
-function generateLogs(text, damage, playerHP) {
-    const fightTime = getTime();
-    const message = text.replace('[time]', fightTime);
-    let el;
-    if (damage) {
-        el = `<p>${fightTime} ${message} -${damage} ${playerHP}/100</p>`;
-    } else {
-        el = `<p>${fightTime} ${message}</p>`;
-    }
-
-    $chat.insertAdjacentHTML('afterbegin', el);
-}
 
 $fightForm.addEventListener('submit', function(event) {
     event.preventDefault();
@@ -332,14 +42,14 @@ $fightForm.addEventListener('submit', function(event) {
     const enemy = enemyAttack();
     const player = playerAttack();
 
+    // Деструтурируем объекты
     const {enemyHitValue, enemyHit, enemyDefence} = enemy;
     const {playerHitValue, playerHit, playerDefence} = player;
 
     if (playerHit !== enemyDefence) {
         player2.changeHP(playerHitValue);
         player2.renderHP();
-        
-        // const message = getLogMessage('hit', player1.name, player2.name);
+        // Получаем сообщение лога и передаем его для вывода
         generateLogs(getLogMessage('hit', player1.name, player2.name), playerHitValue, player2.hp);
 
     } else {
@@ -349,13 +59,13 @@ $fightForm.addEventListener('submit', function(event) {
     if (playerDefence !== enemyHit) {
         player1.changeHP(enemyHitValue);
         player1.renderHP();
-        // generateLogs('hit', player2.name, player1.name);
+        // Получаем сообщение лога и передаем его для вывода
         generateLogs(getLogMessage('hit', player2.name, player1.name), enemyHitValue, player1.hp);
+
     } else {
-        // generateLogs('defence', player2.name, player1.name);
         generateLogs(getLogMessage('defence', player2.name, player1.name), 0, player1.hp);
     }
-
+    // Показываем результат
     showResult();
 
     console.log('###: p', player);
@@ -364,5 +74,6 @@ $fightForm.addEventListener('submit', function(event) {
 //Вызываем функцию создания игрока
 $arenas.appendChild(createPlayer(player1));
 $arenas.appendChild(createPlayer(player2));
+// Получаем сообщение лога о старте и передаем его для вывода
 generateLogs(getLogMessage('start', player1.name, player2.name));
